refactor: migrate from EditorJS to MonkeysEditor across components
