+++ apps/the_monkeys/src/components/blog/actions/PublishBlogDrawer.tsx
@@ -6,7 +6,6 @@ import { Loader } from '@/components/loader';
 import { BLOG_TOPICS_MAX_COUNT } from '@/constants/topics';
 import { useScheduleState } from '@/hooks/blog/schedule/useScheduleState';
 import useGetAllTopics from '@/hooks/user/useGetAllTopics';
-import { OutputData } from '@editorjs/editorjs';
 import { Button } from '@the-monkeys/ui/atoms/button';
 import {
   Drawer,
@@ -22,6 +21,7 @@ import { Label } from '@the-monkeys/ui/atoms/label';
 import { Skeleton } from '@the-monkeys/ui/atoms/skeleton';
 import { Switch } from '@the-monkeys/ui/atoms/switch';
 import { TextArea } from '@the-monkeys/ui/atoms/text-area';
+import { OutputData } from 'monkeys-editor';
 import { twMerge } from 'tailwind-merge';
 
 import ScheduleDateTimePicker from '../ScheduleDateTimePicker';
@@ -41,7 +41,7 @@ const INVALID_IMAGE_EXTENSIONS = ['gif', 'apng'];
 
 /**
  * Extracts blog preview metadata (title, description, image) from
- * EditorJS output data.
+ * MonkeysEditor output data.
  */
 const extractBlogPreview = (data: OutputData | null) => {
   const title =
