+++ apps/the_monkeys/src/components/editor/customBlocks/CustomListBlock/index.ts
@@ -1,4 +1,4 @@
-import { API, BlockTool, ToolboxConfig } from '@editorjs/editorjs';
+import { API, BlockTool, ToolboxConfig } from 'monkeys-editor';
 
 import './style.css';
 import { ConstructorArgs, ListItemData, ListToolData } from './utils/interface';
