+++ apps/the_monkeys/src/config/editor/monkeys_editor.config.ts
@@ -3,16 +3,16 @@ import CustomList from '@/components/editor/customBlocks/CustomListBlock';
 import CustomEmbed from '@/components/editor/customBlocks/EmbedBlock';
 import axiosInstanceV2 from '@/services/api/axiosInstanceV2';
 import Delimiter from '@editorjs/delimiter';
-import { EditorConfig } from '@editorjs/editorjs';
 import Header from '@editorjs/header';
 import Image from '@editorjs/image';
 import List from '@editorjs/list';
 import Paragraph from '@editorjs/paragraph';
 import Quote from '@editorjs/quote';
 import Table from '@editorjs/table';
+import { EditorConfig } from 'monkeys-editor';
 
 export const getEditorConfig = (blogId: string): EditorConfig => ({
-  holder: 'editorjs_editor-container',
+  holder: 'monkeys_editor_editor-container',
   tools: {
     header: {
       class: Header,
