+++ apps/the_monkeys/src/config/editor/monkeys_editor_readonly.config.ts
@@ -3,16 +3,16 @@ import CustomList from '@/components/editor/customBlocks/CustomListBlock';
 import CustomEmbed from '@/components/editor/customBlocks/EmbedBlock';
 import TitleBlockTool from '@/components/editor/customBlocks/TitleBlock';
 import Delimiter from '@editorjs/delimiter';
-import { EditorConfig } from '@editorjs/editorjs';
 import Header from '@editorjs/header';
 import Image from '@editorjs/image';
 import List from '@editorjs/list';
 import Paragraph from '@editorjs/paragraph';
 import Quote from '@editorjs/quote';
 import Table from '@editorjs/table';
+import { EditorConfig } from 'monkeys-editor';
 
 export const editorConfig: EditorConfig = {
-  holder: 'editorjs_editor-container',
+  holder: 'monkeys_editor_editor-container',
   readOnly: true,
   tools: {
     header: {
